Rentrée 2025

+++ docs/mp2i/files/C1/cast.c
@@ -1,10 +1,14 @@
 #include <stdio.h>
 
-float division(float num, float den){
-    float res = num/den;
-    return res;}
+float division(int num, int den)
+{
+    float res = num / den;
+    return res;
+}
 
-int main(){   
-    float deux_tiers= division(2,3);
-    printf("2/3 = %f\n",deux_tiers);
-    return 0;}
+int main()
+{
+    float deux_tiers = division(2, 3);
+    printf("2/3 = %f\n", deux_tiers);
+    return 0;
+}
